Refactor out EmptyState into dedicated component
