+++ plugins/code-versions/src/components/App.tsx
@@ -1,8 +1,9 @@
 import { framer } from "framer-plugin"
 import { useEffect } from "react"
-import CodeFileView from "./components/CodeFileView"
-import { MutationState, useCodeFileVersions } from "./hooks/useCodeFileVersions"
-import { StatusTypes, useSelectedCodeFile } from "./hooks/useSelectedCodeFile"
+import { MutationState, useCodeFileVersions } from "../hooks/useCodeFileVersions"
+import { StatusTypes, useSelectedCodeFile } from "../hooks/useSelectedCodeFile"
+import CodeFileView from "./CodeFileView"
+import { EmptyState } from "./EmptyState"
 
 export default function App() {
     useEffect(() => {
@@ -121,21 +122,6 @@ export default function App() {
                 <CodeFileView state={state} selectVersion={selectVersion} restoreVersion={restoreVersion} />
             </div>
         </Layout>
-    )
-}
-
-function EmptyState() {
-    return (
-        <div className="flex flex-col items-center justify-center space-y-3 h-screen w-screen animate-(--fade-in-animation)">
-            <img src="/logo.svg" className="rounded-lg" />
-
-            <div className="space-y-2 text-center max-w-36">
-                <h2 className="font-semibold text-framer-text-primary text-xs leading-[1.2]">Code Versions</h2>
-                <p className="text-framer-text-tertiary text-xs leading-[1.5]">
-                    Select a Code Component to restore previous versions.
-                </p>
-            </div>
-        </div>
     )
 }
