+++ plugins/code-versions/src/main.tsx
@@ -3,7 +3,7 @@ import "./styles.css"
 import React from "react"
 import ReactDOM from "react-dom/client"
 
-import App from "./App.tsx"
+import App from "./components/App.tsx"
 
 const root = document.getElementById("root")
 if (!root) throw new Error("Root element not found")
